+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+- Add nonroot variant for the kitchen sink image
+([#277](https://github.com/pulumi/pulumi-docker-containers/pull/277)
+
+- Add per language versions for ubi images
+  ([#260](https://github.com/pulumi/pulumi-docker-containers/pull/260))
+
+## 3.132.0
+
 - Add dotnet 8.0 to the kitchen sink image
   ([#259](https://github.com/pulumi/pulumi-docker-containers/pull/259)